Model for prediction of twitter sentiment using Natural Language Processing
In this project, I have built a model for prediction of twitter sentiment using Natural Language Processing. The Dataset contains the text column containing tweets of people and a 'sentiment' column that contains categorical data that classifies the text into 'Positive', 'Negative' or 'Neutral' sentiment. A model was built and trained using the preprocessed data derived from the original dataset and predictions were made using test data. The results were compared and produced.